Continued alignment with codebase conventions
* removing docstrings in unit tests in favor of more meaningful test names
* refactor libguides and researchdatabases tests into springshare
* make error and warning logging message patterns from transform consistent with other transforms
